chore: Add generate-proto.mjs to tsconfig include
Update tsconfig.json to include the file "generate-proto.mjs" in the "include" section. This change ensures that the "generate-proto.mjs" file is included in the TypeScript compilation process.

feat: Add MUI theme provider to app layout

Import the necessary dependencies and add the MUI ThemeProvider and CssBaseline components to the app layout in layout.tsx. This change enables the usage of MUI theming in the application.

chore: Update npm dependencies

Update the npm dependencies in package.json to the latest stable versions. This includes updates to the following packages:
- @emotion/react@11.11.4
- @emotion/styled@11.11.5
- @grpc/grpc-js@1.10.6
- @mui/material@5.15.15
- @prisma/client@5.12.1
- grpc-web@1.5.0
- protobufjs@7.2.6

chore: Add dev dependencies

Add new dev dependencies to package.json:
- @grpc/proto-loader@0.7.12
- bufferutil@4.0.8
- eslint@8
- eslint-config-next@14.2.1
- grpc-tools@1.12.4
- prisma@5.12.1
- protoc-gen-grpc-web@1.4.2
- ts-node@10.9.2
- ts-protoc-gen@0.15.0
- typescript@5.4.5
- utf-8-validate@6.0.3

chore: Update package.json type

Update the "type" field in package.json to "module". This change indicates that the project is using ECMAScript modules.